Added mandoc_a2time() for proper date conversion. Fitted TH and Dd handlers to use mandoc_a2time(). Documented date syntax for -man, fixed documentation for -mdoc.

+++ html.c
@@ -322,7 +322,8 @@ print_encode(struct html *h, const char *p)
 		sz = strcspn(p, "\\<>&");
 
 		fwrite(p, 1, sz, stdout);
-		p += (int)sz;
+		p += /* LINTED */
+			sz;
 
 		if ('\\' == *p) {
 			print_escape(h, &p);
